fix: error message in case Jackett is not installed locally (#41)
So when I tried to install Jackett and test it on my local computer, the
Jackett app and its Serverconfig file was actually located in
"C:\Programs\Jackett" instead of "C:\Users\<User>\AppData\Local\Jackett"
like in the platformdirs library. It could have potentially caused
problem when you try it on Windows.

This is like my first Good first issue so just let me know if something
is wrong, I'll try my best to fix it.
Cheers.

Resolves #39