feat: search command to run w/o welcome screen (#148)
This pull request introduces a `search` command to the CLI,
enabling users to initiate a search directly from the command line.

Previously, users had to navigate through the welcome screen to start a
search. With this update, they can now run `torrra search <query>` to
bypass the welcome screen and go straight to the search results,
streamlining the user experience.

Key changes:
- Added a `search` command in `src/torrra/__main__.py`.
- Modified `TorrraApp` in `src/torrra/app.py` to accept a `search_query`
and conditionally display the `SearchScreen` on startup.